+++ CHANGELOG.rst
@@ -156,3 +156,4 @@ Author: Nikhileswar Reddy <nreddy@octolabs.io>
 * Use ruff format to format the code (#183) (#190)
 * Use ruff check --fix to make style changes (#183) (#192)
 * Added github actions CI (#191) (#193)
+* Be able to run unit tests on dalton and flowsynth (#182)